src/requirements.txt
aniso8601==7.0.0
bokeh==2.2.3
certifi==2020.6.20
chardet==3.0.4
django==2.2.17
django-environ==0.4.5
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-mathfilters==1.0.0
django-mysql==3.9.0
django-split-settings==1.0.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
jinja2==2.11.2
markupsafe==1.1.1
mysqlclient==2.0.1
numpy==1.19.4
packaging==20.4
pillow==8.0.1
promise==2.3
pyjwt==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pyyaml==5.3.1
requests==2.24.0
rx==1.6.1
sentry-sdk==0.16.5
singledispatch==3.4.0.3
six==1.15.0
sql==0.4.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.7.4.3
unidecode==1.1.1
urllib3==1.25.11
